pyproject.toml
[project]
name = "orbit-setup-script"
version = "0.1.0"
description = "Setup script for Orbit worker and KAI services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-pubsub>=2.29.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "psycopg2-binary>=2.9.10",
]
